manual rebase to add evaluator
